+++ model/AnnuaireGroupeSQL.php
@@ -28,12 +28,16 @@ SQL;
         return $this->queryOne($sql, $id_e, $nom);
     }
 
+    public function getInfoById(int $id_g): array|false
+    {
+        $sql = <<<SQL
+SELECT * FROM annuaire_groupe WHERE id_g=?
+SQL;
+        return $this->queryOne($sql, $id_g);
+    }
+
     public function add(int $id_e, string $nom): int
     {
-        $id_g = $this->getFromNom($id_e, $nom);
-        if ($id_g) {
-            return $id_g;
-        }
         $sql = <<<SQL
 INSERT INTO annuaire_groupe (id_e,nom) VALUES (?,?)
 SQL;
@@ -106,7 +110,7 @@ SQL;
         $this->query($sql, $id_g, $id_a);
     }
 
-    public function deleleteFromAllGroupe(int $id_a): void
+    public function deleteFromAllGroupes(int $id_a): void
     {
         $sql = <<<SQL
 DELETE FROM annuaire_groupe_contact WHERE id_a=?