+++ CHANGELOG.md
@@ -1,12 +1,15 @@
-# [4.1.20] -
+# [4.1.20] - 2026-07-06
 
 ## Évolutions
 
+- Création des services `AnnuaireContactService`, `AnnuaireExportService`, `AnnuaireGroupeService` et `AnnuaireImportService` pour les opérations sur l'annuaire #2478
 - Suppression du paramètre interne `id_e` dans `AnnuaireGroupeSQL`, `id_e` doit être directement passé à l'appel des méthodes #2477
 
 ## Dépréciation
 
 - La classe `AnnuaireGroupe` est dépréciée au profit de `AnnuaireGroupeSQL` #2477
+- La classe `AnnuaireExporter` est dépréciée au profit de `AnnuaireExportService` #2478
+- La classe `AnnuaireImporter` est dépréciée au profit de `AnnuaireImportService` #2478
 
 # [4.1.19] - 2026-06-01